Moving weight decorator to the correct location into Rope code (base.utils)
